Update system done (Not tested)

+++ Physic/forces.c
@@ -23,7 +23,7 @@ struct vector* gra_force(struct item *i1, struct item *i2)
   float G = 6.672 * powf(10,-11);
   float m1 = i1->mass, m2 = i2->size;
   float r = distance(i1,i2);
-  float F_magnitude = (G*m1*m2)/powl(r,2);
+  float F_magnitude = (G*m1*m2)/powf(r,2);
   struct vector *F = sub_vector(&(i2->position), clone_vector(&(i1->position))); 
   F = scalar_product_vector(1/r,F); //finding vector unit
   F = scalar_product_vector(F_magnitude, F);
